+++ src/wordless_text/wordless_text_processing.py
@@ -845,11 +845,9 @@ def wordless_lemmatize(main, tokens, lang,
                 lemmas.append(morphological_analyzer.parse(token)[0].normal_form)
         # Tibetan
         elif lemmatizer == main.tr('botok - Tibetan Lemmatizer'):
-            word_tokenizer = main.settings_custom['word_tokenization']['word_tokenizers'][lang]
-
-            botok_tokenizer = wordless_text_utils.check_botok_tokenizers(main,
-                                                                         word_tokenizer = word_tokenizer)
-            tokens = botok_tokenizer.tokenize(' '.join(tokens))
+            wordless_text_utils.check_word_tokenizers(main,
+                                                      lang = 'bod')
+            tokens = main.botok_word_tokenizer.tokenize(' '.join(tokens))
 
             for token in tokens:
                 if token.lemma: